pretrained darknet19 voc2007 trainval mAP 86.9 voc2007 test 58.4

+++ util/visualize.py
@@ -128,7 +128,7 @@ def plot_boxes(im_data, boxes, gt_classes=None, class_names=None):
     im_data -- image data with boxes
     """
     if isinstance(im_data, torch.Tensor):
-        im_data = im_data.permute(1, 2, 0).numpy().astype(np.uint8)
+        im_data = im_data.permute(1, 2, 0).numpy()
         boxes = boxes.numpy()
         gt_classes = gt_classes.numpy() if gt_classes is not None else None
     elif isinstance(im_data, PIL.JpegImagePlugin.JpegImageFile):
